Changed events package location and moved applyBlockLoop to a callback.
