+++ utils/event_handler.go
@@ -8,8 +8,8 @@ import (
 )
 
 // SubscribeAndHandleEvents subscribes to events and sends back a callback
-func SubscribeAndHandleEvents(ctx context.Context, pubsubServer *pubsub.Server, clientID string, eventQuery pubsub.Query, callback func(event pubsub.Message), logger log.Logger) {
-	subscription, err := pubsubServer.Subscribe(ctx, clientID, eventQuery)
+func SubscribeAndHandleEvents(ctx context.Context, pubsubServer *pubsub.Server, clientID string, eventQuery pubsub.Query, callback func(event pubsub.Message), logger log.Logger, outCapacity ...int) {
+	subscription, err := pubsubServer.Subscribe(ctx, clientID, eventQuery, outCapacity...)
 	if err != nil {
 		logger.Error("failed to subscribe to events")
 		panic(err)
